requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.13.0
elasticsearch==8.13.0
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
idna==3.7
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
